Move extract function to module
